Another try on the builder pattern design execise
