+++ Test/src/main/java/org/example/TestBillBuilder.java
@@ -1,0 +1,11 @@
+package org.example;
+
+public class TestBillBuilder {
+    public static void main(String[] args) {
+        BillDirector director = new BillDirector();
+        BillBuilder builder = new BillBuilder();
+        director.makeBill(builder);
+        Bill bill = builder.getBill();
+        System.out.println(bill.toString());
+    }
+}
